docs: 更新README.md详细目录树，同步HealthDetect项目进度和需求文档分析 (2026-01-09)
- README.md: 更新至16 Parts SDK系列文档，新增详细目录树
- HealthDetect项目重建总结.md: 新增问题36需求文档执行对照分析
- AWRL6844_HealthDetect需求文档v2.md: 同步更新
- cli.c/health_detect_main.h/radar_control.c: L-SDK API修复

+++ project-code/AWRL6844_HealthDetect/src/mss/source/health_detect_main.h
@@ -127,8 +127,8 @@ typedef struct HealthDetect_CliCfg_t
     DetectionZone_t         zones[HEALTH_DETECT_MAX_ZONES]; /**< Detection zones */
 
     /* Channel Configuration */
-    uint8_t                 rxChannelEn;        /**< RX channel enable mask */
-    uint8_t                 txChannelEn;        /**< TX channel enable mask */
+    uint16_t                rxChannelEn;        /**< RX channel enable mask */
+    uint16_t                txChannelEn;        /**< TX channel enable mask */
 
     /* Antenna Configuration */
     uint8_t                 numTxAntennas;      /**< Number of TX antennas */
